+++ packages/contracts/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./account";
 export * from "./auth";
 export * from "./category";
+export * from "./dashboard";
 export * from "./transaction";